backend/requirements.txt
flask
requests
flask-mysql
cerberus
flask_cors